pyproject.toml
[tool.poetry]
name = "radigenius"
version = "1.0.0"
description = ""
authors = ["Alireza Safari <alireza.safaree@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.2.18"
Pillow = "10.3.0"
asgiref = "^3.7.2"
sqlparse = "^0.5.0"
psycopg2-binary = "^2.9.7"
djangorestframework = "^3.14.0"
djangorestframework-camel-case = "^1.4.2"
djangorestframework-simplejwt = "^5.3.0"
python-decouple = "^3.8"
Markdown = "^3.4.4"
django-filter = "^23.2"
drf-spectacular = "^0.26.4"
django-cors-headers = "^4.2.0"
seqlog = "^0.3.29"
gunicorn = "^22.0.0"
django-debug-toolbar = "^4.2.0"
cerberus = "^1.3.5"
python-magic = "^0.4.27"
sentry-sdk = {extras = ["django"], version = "^2.20.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"
black = "^24.3.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
